+++ src/generateRolls/applicators/index.ts
@@ -1,0 +1,6 @@
+export * from './apply-drop'
+export * from './apply-explode'
+export * from './apply-replace'
+export * from './apply-reroll'
+export * from './apply-single-cap'
+export * from './apply-unique'
