+++ src/generateRolls/generate-rolls.ts
@@ -1,7 +1,7 @@
 import { RollDie, RollParameters, RollTotals } from 'types'
 import { sumArray } from 'utils'
 
-import { rollParsers, totalParsers } from './parsers'
+import { applyDrop, applyExplode, applyReplace, applyReroll, applySingleCap, applyUnique } from './applicators'
 
 export function generateRolls(
   rollTotals: RollTotals,
@@ -10,14 +10,40 @@ export function generateRolls(
 ): [number, RollTotals] {
   let modifiedRollTotals = [...rollTotals]
 
-  for (const parser of rollParsers(rollParameters, rollDie)) {
-    modifiedRollTotals = parser([...modifiedRollTotals])
+  const { unique, explode, reroll, plus, minus, cap, sides, rolls, replace, drop } = rollParameters
+
+  if (reroll !== undefined) {
+    modifiedRollTotals = applyReroll(modifiedRollTotals, reroll, rollDie)
+  }
+
+  if (unique !== undefined) {
+    modifiedRollTotals = applyUnique(modifiedRollTotals, { sides, rolls, unique }, rollDie)
+  }
+
+  if (replace !== undefined) {
+    modifiedRollTotals = applyReplace(modifiedRollTotals, replace)
+  }
+
+  if (cap !== undefined) {
+    modifiedRollTotals = modifiedRollTotals.map(applySingleCap(cap))
+  }
+
+  if (drop !== undefined) {
+    modifiedRollTotals = applyDrop(modifiedRollTotals, drop)
+  }
+
+  if (explode !== undefined) {
+    modifiedRollTotals = applyExplode(modifiedRollTotals, { sides }, rollDie)
   }
 
   let modifiedTotal = sumArray([...modifiedRollTotals])
 
-  for (const modifierFunction of totalParsers(rollParameters)) {
-    modifiedTotal = modifierFunction(modifiedTotal)
+  if (plus !== undefined) {
+    modifiedTotal = modifiedTotal + plus
+  }
+
+  if (minus !== undefined) {
+    modifiedTotal = modifiedTotal - Math.abs(minus)
   }
 
   return [modifiedTotal, modifiedRollTotals.map(number => Number(number))]
