--- src/generateRolls/parsers/index.ts
@@ -1,2 +1,0 @@
-export * from './roll-parsers'
-export * from './total-parsers'
