--- src/generateRolls/parsers/parse-cap.ts
@@ -1,9 +1,0 @@
-import { CapOptions, RollTotals } from 'types'
-
-import { parseSingleCap } from './shared/parse-single-cap'
-
-export function parseCapFactory(capOptions: CapOptions<'parameters'>) {
-  return function parseCap(rollTotals: RollTotals) {
-    return rollTotals.map(parseSingleCap(capOptions))
-  }
-}
