--- src/generateRolls/parsers/parse-minus.ts
@@ -1,7 +1,0 @@
-import { RollParameters } from 'types'
-
-export function parseMinusFactory(minus: RollParameters['minus']) {
-  return function parseMinus(temporaryTotal: number) {
-    return minus !== undefined ? temporaryTotal - minus : temporaryTotal
-  }
-}
