--- src/generateRolls/parsers/parse-plus.ts
@@ -1,7 +1,0 @@
-import { RollParameters } from 'types'
-
-export function parsePlusFactory(plus: RollParameters['plus']) {
-  return function parsePlus(temporaryTotal: number) {
-    return plus !== undefined ? temporaryTotal + plus : temporaryTotal
-  }
-}
